requirements.txt
coinmarketcapapi
matplotlib
numpy
praw
python_coinmarketcap
python_telegram-bot
urllib3
datetime
pandas
psycopg2-binary
sqlalchemy
english-words